requirements.txt
django>=5.0.0,<5.1.2
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers